+++ Wily-ios/WilyTests/WilyYouTubeURLTests.m
@@ -1,0 +1,21 @@
+@import XCTest;
+#import "WilyYouTubeURL.h"
+
+@interface WilyYouTubeURLTests : XCTestCase
+
+@end
+
+@implementation WilyYouTubeURLTests
+
+- (void)testExtraction {
+  XCTAssertEqualObjects(@"123", WilyYouTubeSongFromURLString(@"http://youtu.be/123")[@"id"]);
+  XCTAssertEqualObjects(@"123", WilyYouTubeSongFromURLString(@"http://youtube.com/v/123")[@"id"]);
+  XCTAssertNil(WilyYouTubeSongFromURLString(@"http://youtube.com/z/123"));
+}
+
+- (void)testConstruction {
+  NSDictionary *song = @{@"id": @"123"};
+  XCTAssertEqualObjects(@"http://youtu.be/123", WilyYouTubeURLStringFromSong(song));
+}
+
+@end